requirements.txt
# Kubernetes / Kubeflow Pipelines SDK
kfp==2.7.0

# Vertex AI SDK
google-cloud-aiplatform==1.42.1

# YAML (caso use configs por ambiente)
PyYAML==6.0.1

